requirements.txt
asgiref==3.9.1
Django==5.2.6
django-filter==25.1
django-phonenumber-field==8.1.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
phonenumbers==9.0.13
phonenumberslite==9.0.13
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
